Ads added

+++ app/memory/useStorage.js
@@ -39,7 +39,7 @@ const removeData = async (dataKeys, setStorageUsed, storageUsed) => {
   try {
     asyncForEach(keys, async (key) => {
       const data = await getData(key);
-      FileSystem.deleteAsync(data[keyfields.THUMB]);
+      if (data[keyfields.THUMB]) FileSystem.deleteAsync(data[keyfields.THUMB]);
     });
     await AsyncStorage.multiRemove(keys);
     setStorageUsed(!storageUsed);
